_archive_scripts/pyproject.toml
[project]
name = "dgn"
version = "0.1.0"
authors=[
    {name="Archie Dobson"},
]
description="Diffusion Graph Nets for Computational Fluid Dynamics applications"
readme="README.md"
requires-python="==3.11.14"
dependencies=[
    "tqdm",
    "numpy",
    "scipy",
    "h5py",
    "matplotlib",
    "h5py",
    "torch_geometric",
    "tensorboard",
    "jupyterlab",
    "vtk",
    "bayesian-torch",
    "pot",
    "torch_cluster",
    "dgn4cfd",
]

[project.optional-dependencies]
doc = [
    "sphinx",
]

[tool.uv.extra-build-dependencies]
torch-cluster = ["torch"]

[tool.uv.workspace]
members = [
    "dgn_move",
]

[tool.uv.sources]
dgn4cfd = { workspace = true, editable = true }
